+++ scipy/stats/tests/test_discrete_basic.py
@@ -8,6 +8,8 @@ import nose
 from scipy.lib.six import xrange
 
 from scipy import stats
+from common_tests import (check_normalization, check_moment, check_mean_expect,
+        check_var_expect, check_skew, check_kurt)
 
 DECIMAL_meanvar = 0  # 1  # was 0
 
@@ -357,75 +359,6 @@ def check_named_args(distfn, x, shape_args, defaults, meths):
 def check_scale_docstring(distfn):
     npt.assert_('scale' not in distfn.__doc__)
 
-
-## these are copy-pasted from test_continuous_basic, verbatim.
-def check_normalization(distfn, args, distname):
-    norm_moment = distfn.moment(0, *args)
-    npt.assert_allclose(norm_moment, 1.0)
-
-    # this is a temporary plug: either ncf or expect is problematic; 
-	# best be marked as a knownfail, but I've no clue how to do it.
-    if distname == "ncf":
-        atol, rtol = 1e-5, 0
-    else:
-        atol, rtol = 1e-7, 1e-7
-
-    norm_expect = distfn.expect(lambda x: 1, args=args)
-    npt.assert_allclose(norm_expect, 1.0, atol=atol, rtol=rtol,
-            err_msg=distname, verbose=True)
-
-    norm_cdf = distfn.cdf(distfn.b, *args)
-    npt.assert_allclose(norm_cdf, 1.0)
-
-
-def check_moment(distfn, arg, m, v, msg):
-    m1 = distfn.moment(1, *arg)
-    m2 = distfn.moment(2, *arg)
-    if not np.isinf(m):
-        npt.assert_almost_equal(m1, m, decimal=10, err_msg=msg +
-                            ' - 1st moment')
-    else:                     # or np.isnan(m1),
-        npt.assert_(np.isinf(m1),
-               msg + ' - 1st moment -infinite, m1=%s' % str(m1))
-
-    if not np.isinf(v):
-        npt.assert_almost_equal(m2 - m1 * m1, v, decimal=10, err_msg=msg +
-                            ' - 2ndt moment')
-    else:                     # or np.isnan(m2),
-        npt.assert_(np.isinf(m2),
-               msg + ' - 2nd moment -infinite, m2=%s' % str(m2))
-
-
-def check_mean_expect(distfn, arg, m, msg):
-    if np.isfinite(m):
-        m1 = distfn.expect(lambda x: x, arg)
-        npt.assert_almost_equal(m1, m, decimal=5, err_msg=msg +
-                            ' - 1st moment (expect)')
-
-
-def check_var_expect(distfn, arg, m, v, msg):
-    if np.isfinite(v):
-        m2 = distfn.expect(lambda x: x*x, arg)
-        npt.assert_almost_equal(m2, v + m*m, decimal=5, err_msg=msg +
-                            ' - 2st moment (expect)')
-
-
-def check_skew(distfn, arg, m, v, s, msg):
-    if np.isfinite(s):
-        m3e = distfn.expect(lambda x: np.power(x-m, 3), arg)
-        npt.assert_almost_equal(m3e, s * np.power(v, 1.5), 
-                decimal=5, err_msg=msg + ' - skew')
-    else:
-        npt.assert_(np.isnan(s))
-
-
-def check_kurt(distfn, arg, m, v, k, msg):
-    if np.isfinite(k):
-        m4e = distfn.expect(lambda x: np.power(x-m, 4), arg)
-        npt.assert_almost_equal(m4e, (k + 3.) * np.power(v, 2), 
-                decimal=5, err_msg=msg + ' - kurtosis')
-    else:
-        npt.assert_(np.isnan(k))
 
 if __name__ == "__main__":
     # nose.run(argv=['', __file__])
